refactor(lint): modularize custom rules
- Group custom rules by policy domain under the lint tool package
- Extract shared path and Vue AST traversal support
- Add direct rule regressions for conditional spreads and AST-backed checks
