+++ tools/lint/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@open-pencil/lint-tools",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "test": "bun test tests"
+  }
+}
