+++ tools/lint/src/rules/vue/index.js
@@ -1,0 +1,17 @@
+export { noBrowserSideEffectsInVue, noDocumentQuerySelectorInVue } from './browser.js'
+export { noLargePropertySectionComponents } from './size.js'
+export {
+  noDynamicDataTestIdInVue,
+  noGeneratedTestIdLiterals,
+  noInvalidTestIdAttributes,
+  noRawTestIdSelectorsInTests,
+  noRawTestIdStringProps,
+  noTestIdHelperBindInVue
+} from './test-ids.js'
+export {
+  noHardcodedTipLabelsInVue,
+  noNativeTitleAttributesInVue,
+  noRawSvgInAppVueTemplates,
+  noUiHelperCallsInVueTemplates,
+  noVueStyleBlocks
+} from './template.js'
